Creates a basic structure for streaming Kafka jobs
Each of the subdirectories in this project are effectively a copy of
each other with the content of a single Job.scala changed. With this
change, we should have a common KafkaStreamingApplication that can be
shared between jobs. The only side-effect of this change is that ALL
Spark jobs are now in the SAME JAR file and therefore must be invoked
with different main objects.

Change-Id: Ib167e281143ff93cbe4da35e34bda7d6a563d58a