Moves the original effective jobs processing into the new core
Some adaptation has been made to the original (left in place for
study, for now) to remove the use of cartesian and replace it with
DStream.join using a paired (K, V) dstream. The concepts of effective
and envelope have been encapsulated into classes that will implement
the filtering from the original job.

Change-Id: I8ef78d58b0f1bbec4eae755e17d2417c05be39a7