Removes "envelope" concepts from the Effective job
Previously, we preserved meta-information about inbound documents in a
Cassandra table called "envelopes". We used this table as the driver
for the effective job (based on the document_id). This was unnecessary
and led to preserving information we shouldn't preserve.

Now, everything we need to fuel the process, per-document, is sent on
the Kafka queue from the schedule service.